Update top bar positioning
